replace mux router with net/http router

+++ api/server.go
@@ -5,14 +5,14 @@ import (
 	"employees/internal/models"
 	"encoding/json"
 	"net/http"
+	"strconv"
 
-	"github.com/gorilla/mux"
 	"go.uber.org/zap"
 )
 
 type Server struct {
 	logger     *zap.Logger
-	router     *mux.Router
+	router     *http.ServeMux
 	listenAddr string
 	db         db.Database
 }
@@ -21,29 +21,55 @@ func NewServer(listenAddr string, logger *zap.Logger, db db.Database) *Server {
 	return &Server{
 		logger:     logger,
 		listenAddr: listenAddr,
-		router:     mux.NewRouter(),
+		router:     http.NewServeMux(),
 		db:         db,
 	}
 }
 
 func (s *Server) Start() error {
-	s.router.HandleFunc("/employee", s.handleCreateEmployee).Methods("POST")
-	s.router.HandleFunc("/employee/{id}", s.handleGetEmployee).Methods("GET")
-	s.router.HandleFunc("/employee/{id}", s.handleUpdateEmployee).Methods("PUT")
-	s.router.HandleFunc("/employee/{id}", s.handleDeleteEmployee).Methods("DELETE")
-
+	s.router.HandleFunc("/employee", s.handleEmployee)
 	return http.ListenAndServe(s.listenAddr, s.router)
+}
+
+func (s *Server) handleEmployee(w http.ResponseWriter, r *http.Request) {
+	switch r.Method {
+	case "POST":
+		s.handleCreateEmployee(w, r)
+	case "GET":
+		s.handleGetEmployee(w, r)
+	case "PUT":
+		s.handleUpdateEmployee(w, r)
+	case "DELETE":
+		s.handleDeleteEmployee(w, r)
+	}
 }
 
 func (s *Server) handleCreateEmployee(w http.ResponseWriter, r *http.Request) {
 	var emp models.Employee
 	if err := json.NewDecoder(r.Body).Decode(&emp); err != nil {
-		s.logger.Error("Invalid request body", zap.Error(err))
+		s.logger.Error("Employee creation failed", zap.Error(err))
 		http.Error(w, "Invalid request body", http.StatusBadRequest)
 		return
 	}
 
-	err := s.db.CreateEmployee(&emp)
+	maxID, err := s.db.GetMaxEmployeeID()
+	if err != nil {
+		s.logger.Error("Failed to get max employee ID", zap.Error(err))
+		http.Error(w, "Employee creation failed", http.StatusInternalServerError)
+		return
+	}
+
+	nextId, err := strconv.Atoi(maxID)
+	if err != nil {
+		s.logger.Error("Failed to convert max employee ID to int", zap.Error(err))
+		http.Error(w, "Employee creation failed", http.StatusInternalServerError)
+		return
+	}
+
+	nextId += 1
+	emp.ID = strconv.Itoa(nextId)
+
+	err = s.db.CreateEmployee(&emp)
 	if err != nil {
 		s.logger.Error("Employee creation failed", zap.Error(err))
 		http.Error(w, "Failed to create employee", http.StatusBadRequest)
@@ -59,8 +85,13 @@ func (s *Server) handleCreateEmployee(w http.ResponseWriter, r *http.Request) {
 }
 
 func (s *Server) handleGetEmployee(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	id := vars["id"]
+	id := r.URL.Query().Get("id")
+
+	if id == "" {
+		http.Error(w, "ID is required", http.StatusBadRequest)
+		s.logger.Error("ID is required")
+		return
+	}
 
 	emp, err := s.db.GetEmployee(id)
 	if err != nil {
@@ -77,8 +108,19 @@ func (s *Server) handleGetEmployee(w http.ResponseWriter, r *http.Request) {
 }
 
 func (s *Server) handleUpdateEmployee(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	id := vars["id"]
+	id := r.URL.Query().Get("id")
+
+	if id == "" {
+		http.Error(w, "ID is required", http.StatusBadRequest)
+		s.logger.Error("ID is required")
+		return
+	}
+
+	if !validateId(id) {
+		http.Error(w, "Invalid ID", http.StatusBadRequest)
+		s.logger.Error("Invalid ID")
+		return
+	}
 
 	var emp models.Employee
 	if err := json.NewDecoder(r.Body).Decode(&emp); err != nil {
@@ -104,8 +146,13 @@ func (s *Server) handleUpdateEmployee(w http.ResponseWriter, r *http.Request) {
 }
 
 func (s *Server) handleDeleteEmployee(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	id := vars["id"]
+	id := r.URL.Query().Get("id")
+
+	if id == "" {
+		http.Error(w, "ID is required", http.StatusBadRequest)
+		s.logger.Error("ID is required")
+		return
+	}
 
 	if err := s.db.DeleteEmployee(id); err != nil {
 		http.Error(w, err.Error(), http.StatusNotFound)
@@ -116,4 +163,11 @@ func (s *Server) handleDeleteEmployee(w http.ResponseWriter, r *http.Request) {
 	s.logger.Info("Employee deleted", zap.Any("employeeId", id))
 
 	w.WriteHeader(http.StatusNoContent)
+}
+
+func validateId(id string) bool {
+	if _, err := strconv.Atoi(id); err != nil {
+		return false
+	}
+	return true
 }
